Transform galleries.rb script into Jekyll plugin.
This means that now all the galleries are automatically generated with each 'git push'. So when adding or removing a photo from a gallery, all that's required is to insert or remove said photo from the filesystem, commit the file, then 'git push', and the file appears in the site, automagically. I'm quite pleased with this.